Created ImageInformationRequest_v3_Test.java
Former-commit-id: 02a221cf3b90d6e56d83f2e5a0e11514a837c435

+++ cineast-core/src/main/java/org/vitrivr/cineast/core/iiif/imageapi/ImageInformation.java
@@ -55,14 +55,14 @@ public interface ImageInformation {
     @Getter
     @Setter
     @JsonProperty
-    public Integer width;
+    public Long width;
 
     @Getter
     @Setter
     @JsonProperty
-    public Integer height;
+    public Long height;
 
-    public SizesItem(int width, int height) {
+    public SizesItem(long width, long height) {
       this.width = width;
       this.height = height;
     }
